LayoutTests/media/media-source/media-source-changetype-tracks.html

<!DOCTYPE html><!-- webkit-test-runner [ SourceBufferChangeTypeEnabled=true ] -->
<html>
<head>
    <title>media-source-changetype-tracks</title>
    <script src="media-source-loader.js"></script>
    <script src="../video-test.js"></script>
    <script>
    var mp4Loader;
    var webmLoader;
    var source;
    var sourceBuffer;
    var i;

    function loaderPromise(loader) {
        return new Promise((resolve, reject) => {
            loader.onload = resolve;
            loader.onerror = reject;
        });
    }

    window.addEventListener('load', async event => {
        try {
            findMediaElement();

            mp4Loader = new MediaSourceLoader('content/test-fragmented-manifest.json');
            await loaderPromise(mp4Loader);
            webmLoader = new MediaSourceLoader('content/test-vp8-24fps-manifest.json');
            await loaderPromise(webmLoader);

            source = new MediaSource();
            run('video.src = URL.createObjectURL(source)');
            await waitFor(source, 'sourceopen');
            waitForEventAndFail('error');

            run(`sourceBuffer = source.addSourceBuffer('${mp4Loader.type()}')`);
            run('sourceBuffer.appendBuffer(mp4Loader.initSegment())');
            await waitFor(sourceBuffer, 'update');

            consoleWrite('<br/>Append some MP4 media segments');
            for (i = 0; i < 2; i++) {
              run(`sourceBuffer.appendBuffer(mp4Loader.mediaSegment(${i}))`);
              await waitFor(sourceBuffer, 'update');
            }

            consoleWrite('<br/>Change type to WebM');
            run(`sourceBuffer.changeType('${webmLoader.type()}')`);
            run(`sourceBuffer.timestampOffset = mp4Loader.mediaSegmentEndTime(${i - 1})`);
            run('sourceBuffer.appendBuffer(webmLoader.initSegment())');
            await waitFor(sourceBuffer, 'update');

            consoleWrite('<br/>Append WebM media segments and play');
            for (i = 0; i < webmLoader.mediaSegmentsLength(); i++) {
              run(`sourceBuffer.appendBuffer(webmLoader.mediaSegment(${i}))`);
              await waitFor(sourceBuffer, 'update');
            }

            run('source.endOfStream()');
            await waitFor(source, "sourceended");

            run('video.play()');
            await waitFor(video, "play");
            await waitFor(video, "ended");

            endTest();
        } catch (e) {
            failTest(`Caught exception: "${e}"`);
        }
    });
    </script>
</head>
<body>
    <div>
        This tests that changeType() works when playing media with more than one track.
    </div>
    <video controls></video>
</body>
</html>
